requirements.txt
colorama==0.4.6
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0